pipeline/webhook_receiver/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
google-cloud-pubsub==2.21.1
structlog==24.1.0
python-dotenv==1.0.1
pydantic>=2.0
